env_virtual_requirements.txt
Adafruit-Blinka==8.53.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-fingerprint==2.2.19
adafruit-circuitpython-requests==4.1.9
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
colorzero==2.0
future==1.0.0
gpiozero==2.0.1
iso8601==2.1.0
machine==0.0.1
mysql-connector==2.2.9
pillow==11.1.0
pycrypto==2.6.1
pyfingerprint==1.5
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
PyYAML==6.0.2
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
smbus==1.1.post2
sysv-ipc==1.1.0
typing_extensions==4.12.2
